example files / script for building python wheel
The python wheel can be built using the `build_wheel.sh` script, called
from within the src folder.  It can then be installed using `pip3
install <wheel_filename>`.  This will create the `artisan` script in
either the `~/.local/bin` folder, or in the bin folder of the active
python venv.

Ideally there should only be one single `requirements.txt` file in the
repository (instead of 8 or so).